refactor: simplify auth core API
