+++ src/application/optional.ts
@@ -1,0 +1,10 @@
+export function optionalProp<Key extends string, Value>(
+  key: Key,
+  value: Value | undefined,
+): { readonly [Property in Key]?: Value } {
+  if (value === undefined) {
+    return {}
+  }
+
+  return { [key]: value } as { readonly [Property in Key]?: Value }
+}
