+++ src/index.ts
@@ -2,8 +2,8 @@ export * from './attribution.js'
 export * from './application/auth-service.js'
 export * from './application/policy.js'
 export * from './domain/types.js'
-export * from './errors'
-export * from './ports'
+export * from './errors.js'
+export * from './ports.js'
 export * from './utils/ids.js'
 export * from './utils/normalization.js'
 export * from './utils/secrets.js'
