+++ src/utils/secrets.ts
@@ -1,6 +1,12 @@
-import { createHash, randomBytes, randomInt, timingSafeEqual } from 'node:crypto'
+import { createHash, createHmac, randomBytes, randomInt, timingSafeEqual } from 'node:crypto'
 
 const SECRET_HASH_PREFIX = 'sha256:'
+const HMAC_SECRET_HASH_PREFIX = 'hmac-sha256:'
+
+export interface SecretHasher {
+  hash(secret: string): string | Promise<string>
+  verify(secret: string, secretHash: string): boolean | Promise<boolean>
+}
 
 export function generateSecret(byteLength = 32): string {
   return randomBytes(byteLength).toString('base64url')
@@ -15,12 +21,37 @@ export function hashSecret(secret: string): string {
 }
 
 export function verifySecret(secret: string, secretHash: string): boolean {
-  if (!secretHash.startsWith(SECRET_HASH_PREFIX)) {
+  return verifyPrefixedSecret(secretHash, hashSecret(secret), SECRET_HASH_PREFIX)
+}
+
+export const sha256SecretHasher: SecretHasher = {
+  hash: hashSecret,
+  verify: verifySecret,
+}
+
+export function createHmacSecretHasher(input: { readonly pepper: string }): SecretHasher {
+  if (!input.pepper) {
+    throw new Error('Secret hasher pepper is required.')
+  }
+
+  const hash = (secret: string): string =>
+    `${HMAC_SECRET_HASH_PREFIX}${createHmac('sha256', input.pepper).update(secret).digest('hex')}`
+
+  return {
+    hash,
+    verify(secret, secretHash): boolean {
+      return verifyPrefixedSecret(secretHash, hash(secret), HMAC_SECRET_HASH_PREFIX)
+    },
+  }
+}
+
+function verifyPrefixedSecret(secretHash: string, actualHash: string, prefix: string): boolean {
+  if (!secretHash.startsWith(prefix)) {
     return false
   }
 
   const expected = Buffer.from(secretHash)
-  const actual = Buffer.from(hashSecret(secret))
+  const actual = Buffer.from(actualHash)
 
   if (actual.length !== expected.length) {
     return false
